PID slightly tuned, but underlying issues probably
The PID is closer to tuned, but it is not even close to reaching it's target velocity. This could be attributed to the strange kinematics issues we have been having. It also depends on the direction, as moving in a diagonal direction seems to make it go faster, as if it was a vector that is not normalized. Point is, it almost works on the surface, but there are probably problems underneath that need to be fixed